+++ CHANGELOG.md
@@ -27,6 +27,7 @@ feat: add `ADMIN_TABLE_DELETE_ROW` and `ADMIN_TABLE_INSERT_ROW` permissions
 
 - feat: add `ADMIN_TABLE_DELETE_ROW` and `ADMIN_TABLE_INSERT_ROW` permissions
 - feat: add deletable and insertable as properties returned by AdminModel::knowledge
+- feat: add TracyBar PDO error status for cases where a PDOException isn’t caught by SeablastPdo, allowing index.php to trigger it directly.
 
 ### Changed